pyproject.toml
[project]
name = "pocket-rag"
version = "0.1.0"
description = "Add your description here"
authors = [
    { name = "shunichi0402", email = "shunichi_nikoniko@yahoo.co.jp" }
]
dependencies = [
    "sqlite-vec>=0.1.6",
    "numpy>=2.2.5",
    "torch>=2.7.0",
    "transformers>=4.51.3",
    "openai>=1.78.0",
    "python-dotenv>=1.1.0",
    "sentencepiece>=0.2.0",
    "pdf2md @ git+https://github.com/shunichi0402/pdf2md.git",
    "mistune>=3.1.3",
]
readme = "README.md"
requires-python = ">= 3.8"

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.rye]
managed = true
dev-dependencies = []

[tool.hatch.metadata]
allow-direct-references = true

[tool.hatch.build.targets.wheel]
packages = ["src/pocket_rag"]
